details view extension is divided into 2 files
