+++ Productivity_Timer/Productivity Timer/Activities/Extensions/A.DetailedVCExtension.swift
@@ -9,96 +9,86 @@ extension ActivityDetailedViewController: UITableViewDelegate, UITableViewDataSo
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 4
+        return 5
     }
 
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-
-        let selectedIndexPath = tableView.indexPathForSelectedRow
-        guard selectedIndexPath?.section != 1 || selectedIndexPath?.section != 3 else { return }
-
-        if selectedIndexPath?.row == 1 {
-            let titleRowEditAction = UIAlertController(title: "Edit Title", message: "Please edit the title", preferredStyle: .alert)
-            titleRowEditAction.addTextField(configurationHandler: { (newTitle) -> Void in
-
-                newTitle.text = self.activity.title
-            })
-
-            let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) -> Void in
-
-                self.presentingViewController?.dismiss(animated: true, completion: nil)
-            })
-
-            let okayAction = UIAlertAction(title: "OK", style: .default, handler: { (action) -> Void in
-
-                SelectedActivity.selectedActivity = self.activity
-
-                self.activity.title = (titleRowEditAction.textFields?.first?.text)!
-
-                guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
-                let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
-
-                if SelectedActivity.selectedActivity != nil {
-//                    guard let editedActivity = SelectedActivity.selectedActivity else { return }
-//                    editedActivity.id = ActivitiesObject.arrayOfActivities.count as NSNumber
-//                    editedActivity.title = self.activity.title
-//                    editedActivity.fav = false
-
-                    do {
-                        try context.save()
-//                        ActivitiesObject.arrayOfActivities.append(editedActivity)
-                        SelectedActivity.selectedActivity = nil
-                    } catch {
-                        print("Can't save the context")
-                    }
-
-                    //                } else  if selectedIndexPath?.row == 3 {
-                    //                    let descRowEditAction = UIAlertController(title: "Edit Description", message: "Please edit the description", preferredStyle: .alert)
-                    //
-                    //                    descRowEditAction.addTextField(configurationHandler: { (newDescription) -> Void in
-                    //                        newDescription.text = self.activity.desc
-                    //                    })
-                    //
-                    //                    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) -> Void in
-                    //
-                    //                        self.presentingViewController?.dismiss(animated: true, completion: nil)
-                    //                    })
-
-                    //                    let okayAction = UIAlertAction(title: "OK", style: .default, handler: { (action) -> Void in
-                    //
-                    //                        SelectedActivity.selectedActivity = ActivitiesObject.arrayOfActivities[indexPath.row]
-                    //
-                    //                        self.activity.desc = (descRowEditAction.textFields?.first?.text)!
-                    //
-                    //                        guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
-                    //
-                    //                        let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
-                    //
-                    //                        if SelectedActivity.selectedActivity != nil {
-                    //                            guard let newActivity = SelectedActivity.selectedActivity else { return }
-                    //                            newActivity.id = ActivitiesObject.arrayOfActivities.count as NSNumber
-                    //                            newActivity.title = self.activity.desc
-                    //                            newActivity.fav = false
-                    //
-                    //                            do {
-                    //                                try context.save()
-                    //
-                    //                                ActivitiesObject.arrayOfActivities.append(newActivity)
-                    //                                SelectedActivity.selectedActivity = nil
-                    //                            } catch {
-                    //                                print("Can't save the context")
-                    //                            }
-
-                }
-
-                tableView.reloadData()
-                self.presentingViewController?.dismiss(animated: true, completion: nil)
-            })
-            titleRowEditAction.addAction(okayAction)
-            titleRowEditAction.addAction(cancelAction)
-            self.present(titleRowEditAction, animated: true, completion: nil)
-        }
-    }
+//    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+//
+//        let selectedIndexPath = tableView.indexPathForSelectedRow
+//        guard selectedIndexPath?.section != 1 || selectedIndexPath?.section != 3 else { return }
+//
+//        if selectedIndexPath?.row == 1 {
+//            let titleRowEditAction = UIAlertController(title: "Edit Title", message: "Please edit the title", preferredStyle: .alert)
+//            titleRowEditAction.addTextField(configurationHandler: { (newTitle) -> Void in
+//                newTitle.text = self.activity.title
+//            })
+//
+//            let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) -> Void in
+//                self.presentingViewController?.dismiss(animated: true, completion: nil)
+//            })
+//
+//            let okayAction = UIAlertAction(title: "OK", style: .default, handler: { (action) -> Void in
+//
+//                SelectedActivity.selectedActivity = self.activity
+//
+//                self.activity.title = (titleRowEditAction.textFields?.first?.text)!
+//
+//                guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
+//                let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
+//
+//                if SelectedActivity.selectedActivity != nil {
+//
+//                    do {
+//                        try context.save()
+//                        SelectedActivity.selectedActivity = nil
+//                    } catch {
+//                        print("Can't save the context")
+//                    }
+//                }
+//
+//                tableView.reloadData()
+//                self.presentingViewController?.dismiss(animated: true, completion: nil)
+//            })
+//            titleRowEditAction.addAction(okayAction)
+//            titleRowEditAction.addAction(cancelAction)
+//            self.present(titleRowEditAction, animated: true, completion: nil)
+//        }
+//
+//        if selectedIndexPath?.row == 3 {
+//            let descRowEditAction = UIAlertController(title: "Edit Description", message: "Please edit the description", preferredStyle: .alert)
+//            descRowEditAction.addTextField(configurationHandler: { (newDescription) -> Void in
+//                newDescription.text = self.activity.desc
+//            })
+//
+//            let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) -> Void in
+//                self.presentingViewController?.dismiss(animated: true, completion: nil)
+//            })
+//
+//            let okayAction = UIAlertAction(title: "OK", style: .default, handler: { (action) -> Void in
+//
+//                SelectedActivity.selectedActivity = self.activity
+//                self.activity.desc = (descRowEditAction.textFields?.first?.text)!
+//
+//                guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
+//                let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
+//
+//                if SelectedActivity.selectedActivity != nil {
+//                    do {
+//                        try context.save()
+//                        SelectedActivity.selectedActivity = nil
+//                    } catch {
+//                        print("Can't save the context")
+//                    }
+//                }
+//
+//                tableView.reloadData()
+//                self.presentingViewController?.dismiss(animated: true, completion: nil)
+//            })
+//            descRowEditAction.addAction(okayAction)
+//            descRowEditAction.addAction(cancelAction)
+//            self.present(descRowEditAction, animated: true, completion: nil)
+//        }
+//    }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let index = indexPath.row
@@ -136,14 +126,20 @@ extension ActivityDetailedViewController: UITableViewDelegate, UITableViewDataSo
             cell.selectionStyle = .none
             cell.backgroundColor = darkMoonColor
             return cell
-            //        case 3:
-            //            let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-            //            if activity.isFavourite {
-            //            cell.imageView?.image = "heart"
-            //            }
-            //            cell.textLabel?.numberOfLines = 0
-            //            cell.selecti onStyle = .none
-            //            return cell
+        case 4:
+            let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
+            cell.selectionStyle = .none
+            cell.backgroundColor = darkMoonColor
+
+            if activity.fav {
+                cell.imageView?.image = UIImage(systemName: "heart.fill")
+                cell.imageView?.tintColor = .systemRed
+            } else {
+                cell.imageView?.image = UIImage(systemName: "heart")
+                cell.imageView?.tintColor = .systemGray
+            }
+
+            return cell
         default:
             fatalError()
         }
